HV-1389 Clean up @CandidateForTck tests
