+++ engine/src/test/java/org/hibernate/validator/test/internal/engine/cascaded/ExtendedChecks1.java
@@ -1,0 +1,11 @@
+/*
+ * Hibernate Validator, declare and validate application constraints
+ *
+ * License: Apache License, Version 2.0
+ * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
+ */
+package org.hibernate.validator.test.internal.engine.cascaded;
+
+
+public interface ExtendedChecks1 {
+}
