--- engine/src/test/java/org/hibernate/validator/test/internal/engine/cascaded/GuavaOptionalValueExtractor.java
@@ -1,23 +1,0 @@
-/*
- * Hibernate Validator, declare and validate application constraints
- *
- * License: Apache License, Version 2.0
- * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
- */
-package org.hibernate.validator.test.internal.engine.cascaded;
-
-import javax.validation.valueextraction.ExtractedValue;
-import javax.validation.valueextraction.ValueExtractor;
-
-import com.google.common.base.Optional;
-
-/**
- * @author Gunnar Morling
- */
-public class GuavaOptionalValueExtractor implements ValueExtractor<Optional<@ExtractedValue ?>> {
-
-	@Override
-	public void extractValues(Optional<@ExtractedValue ?> originalValue, ValueExtractor.ValueReceiver receiver) {
-		receiver.value( null, originalValue.isPresent() ? originalValue.get() : null );
-	}
-}
