+++ engine/src/test/java/org/hibernate/validator/test/internal/engine/typeannotationconstraint/TypeAnnotationDefinedOnAGenericTypeArgumentTest.java
@@ -17,7 +17,6 @@ import javax.validation.constraints.Size;
 
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.testutil.TestForIssue;
-import org.hibernate.validator.testutils.CandidateForTck;
 import org.hibernate.validator.testutils.ValidatorUtil;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
@@ -26,7 +25,6 @@ import org.testng.annotations.Test;
  * @author Guillaume Smet
  */
 @TestForIssue(jiraKey = "HV-1279")
-@CandidateForTck
 public class TypeAnnotationDefinedOnAGenericTypeArgumentTest {
 
 	private Validator validator;
