+++ engine/src/test/java/org/hibernate/validator/test/internal/engine/valueextraction/CascadedOptionalTest.java
@@ -22,8 +22,6 @@ import javax.validation.Validator;
 import javax.validation.constraints.NotNull;
 
 import org.hibernate.validator.testutil.TestForIssue;
-import org.hibernate.validator.testutils.CandidateForTck;
-
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
@@ -33,7 +31,6 @@ import org.testng.annotations.Test;
  * @author Gunnar Morling
  */
 @TestForIssue(jiraKey = "HV-1080")
-@CandidateForTck
 public class CascadedOptionalTest {
 
 	private Validator validator;
