+++ engine/src/test/java/org/hibernate/validator/test/internal/engine/valueextraction/OptionalTypeAnnotationConstraintOnFieldTest.java
@@ -37,8 +37,6 @@ import org.hibernate.validator.constraints.CompositionType;
 import org.hibernate.validator.constraints.ConstraintComposition;
 import org.hibernate.validator.constraints.NotBlank;
 import org.hibernate.validator.testutil.TestForIssue;
-import org.hibernate.validator.testutils.CandidateForTck;
-
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
@@ -48,7 +46,6 @@ import org.testng.annotations.Test;
  * @author Davide D'Alto
  */
 @TestForIssue(jiraKey = "HV-976")
-@CandidateForTck
 public class OptionalTypeAnnotationConstraintOnFieldTest {
 
 	private Validator validator;
