+++ engine/src/test/java/org/hibernate/validator/test/internal/engine/valueextraction/OptionalTypeAnnotationConstraintUsingValidateValueTest.java
@@ -34,8 +34,6 @@ import org.hibernate.validator.constraints.CompositionType;
 import org.hibernate.validator.constraints.ConstraintComposition;
 import org.hibernate.validator.constraints.NotBlank;
 import org.hibernate.validator.testutil.TestForIssue;
-import org.hibernate.validator.testutils.CandidateForTck;
-
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
@@ -44,7 +42,6 @@ import org.testng.annotations.Test;
  *
  * @author George Gastaldi
  */
-@CandidateForTck
 public class OptionalTypeAnnotationConstraintUsingValidateValueTest {
 	private Validator validator;
