--- engine/src/test/resources/org/hibernate/validator/test/internal/xml/containerelementtype/field-canDeclareContainerElementCascades-mapping.xml
@@ -1,23 +1,0 @@
-<constraint-mappings
-        xmlns="http://xmlns.jcp.org/xml/ns/validation/mapping"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/validation/mapping
-            http://xmlns.jcp.org/xml/ns/validation/mapping/validation-mapping-2.0.xsd"
-        version="2.0">
-
-    <default-package>org.hibernate.validator.internal.xml.containerelementtype</default-package>
-
-    <bean class="org.hibernate.validator.test.internal.xml.containerelementtype.ContainerElementTypeConstraintsForFieldXmlMappingTest$FishTank" ignore-annotations="false">
-        <field name="boss">
-            <container-element-type>
-                <valid />
-            </container-element-type>
-        </field>
-    </bean>
-
-    <bean class="org.hibernate.validator.test.internal.xml.containerelementtype.ContainerElementTypeConstraintsForFieldXmlMappingTest$Fish" ignore-annotations="false">
-        <field name="name">
-            <constraint annotation="javax.validation.constraints.NotNull" />
-        </field>
-    </bean>
-</constraint-mappings>
