--- engine/src/test/resources/org/hibernate/validator/test/internal/xml/containerelementtype/getter-canDeclareContainerElementTypeConstraints-mapping.xml
@@ -1,40 +1,0 @@
-<constraint-mappings
-        xmlns="http://xmlns.jcp.org/xml/ns/validation/mapping"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/validation/mapping
-            http://xmlns.jcp.org/xml/ns/validation/mapping/validation-mapping-2.0.xsd"
-        version="2.0">
-
-    <default-package>org.hibernate.validator.internal.xml.containerelementtype</default-package>
-
-    <bean class="org.hibernate.validator.test.internal.xml.containerelementtype.ContainerElementTypeConstraintsForGetterXmlMappingTest$FishTank" ignore-annotations="false">
-        <getter name="model">
-            <container-element-type>
-                <constraint annotation="javax.validation.constraints.Size">
-                    <element name="max">
-                        <value>5</value>
-                    </element>
-                </constraint>
-            </container-element-type>
-        </getter>
-        <getter name="fishCountByType">
-            <container-element-type type-argument-index="0">
-                <constraint annotation="javax.validation.constraints.Size">
-                    <element name="min">
-                        <value>3</value>
-                    </element>
-                    <element name="max">
-                        <value>10</value>
-                    </element>
-                </constraint>
-            </container-element-type>
-            <container-element-type type-argument-index="1">
-                <constraint annotation="javax.validation.constraints.Min">
-                    <element name="value">
-                        <value>1</value>
-                    </element>
-                </constraint>
-            </container-element-type>
-        </getter>
-    </bean>
-</constraint-mappings>
