--- engine/src/test/resources/org/hibernate/validator/test/internal/xml/containerelementtype/getter-canDeclareNestedContainerElementTypeConstraints-mapping.xml
@@ -1,19 +1,0 @@
-<constraint-mappings
-        xmlns="http://xmlns.jcp.org/xml/ns/validation/mapping"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/validation/mapping
-            http://xmlns.jcp.org/xml/ns/validation/mapping/validation-mapping-2.0.xsd"
-        version="2.0">
-
-    <default-package>org.hibernate.validator.internal.xml.containerelementtype</default-package>
-
-    <bean class="org.hibernate.validator.test.internal.xml.containerelementtype.ContainerElementTypeConstraintsForGetterXmlMappingTest$FishTank" ignore-annotations="false">
-        <getter name="fishOfTheMonth">
-            <container-element-type type-argument-index="1">
-                <container-element-type type-argument-index="0">
-                    <constraint annotation="javax.validation.constraints.NotNull" />
-                </container-element-type>
-            </container-element-type>
-        </getter>
-    </bean>
-</constraint-mappings>
