--- engine/src/test/resources/org/hibernate/validator/test/internal/xml/containerelementtype/parameter-canDeclareContainerElementCascades-mapping.xml
@@ -1,25 +1,0 @@
-<constraint-mappings
-        xmlns="http://xmlns.jcp.org/xml/ns/validation/mapping"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/validation/mapping
-            http://xmlns.jcp.org/xml/ns/validation/mapping/validation-mapping-2.0.xsd"
-        version="2.0">
-
-    <default-package>org.hibernate.validator.internal.xml.containerelementtype</default-package>
-
-    <bean class="org.hibernate.validator.test.internal.xml.containerelementtype.ContainerElementTypeConstraintsForParameterXmlMappingTest$IFishTank" ignore-annotations="false">
-        <method name="test4">
-            <parameter type="java.util.Optional">
-                <container-element-type>
-                    <valid />
-                </container-element-type>
-            </parameter>
-        </method>
-    </bean>
-
-    <bean class="org.hibernate.validator.test.internal.xml.containerelementtype.ContainerElementTypeConstraintsForParameterXmlMappingTest$Fish" ignore-annotations="false">
-        <field name="name">
-            <constraint annotation="javax.validation.constraints.NotNull" />
-        </field>
-    </bean>
-</constraint-mappings>
