--- engine/src/test/resources/org/hibernate/validator/test/internal/xml/containerelementtype/returnvalue-canDeclareContainerElementTypeConstraints-mapping.xml
@@ -1,33 +1,0 @@
-<constraint-mappings
-        xmlns="http://xmlns.jcp.org/xml/ns/validation/mapping"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/validation/mapping
-            http://xmlns.jcp.org/xml/ns/validation/mapping/validation-mapping-2.0.xsd"
-        version="2.0">
-
-    <default-package>org.hibernate.validator.internal.xml.containerelementtype</default-package>
-
-    <bean class="org.hibernate.validator.test.internal.xml.containerelementtype.ContainerElementTypeConstraintsForReturnValueXmlMappingTest$IFishTank" ignore-annotations="false">
-        <method name="test1">
-            <return-value>
-                <container-element-type type-argument-index="0">
-                    <constraint annotation="javax.validation.constraints.Size">
-                        <element name="min">
-                            <value>3</value>
-                        </element>
-                        <element name="max">
-                            <value>10</value>
-                        </element>
-                    </constraint>
-                </container-element-type>
-                <container-element-type type-argument-index="1">
-                    <constraint annotation="javax.validation.constraints.Min">
-                        <element name="value">
-                            <value>1</value>
-                        </element>
-                    </constraint>
-                </container-element-type>
-            </return-value>
-        </method>
-    </bean>
-</constraint-mappings>
